anypoint-mcp-guardrails/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anypoint-mcp-guardrails"
version = "0.1.0"
description = "Read-only + scoped-write MCP server for Anypoint Platform (Integrations-NA)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
anypoint-mcp-guardrails = "anypoint_mcp.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["anypoint_mcp*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
